requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httplib2
pydantic
python-decouple
uritemplate

# code style
autoflake8
black
isort
pycodestyle
